Convenience target for bundled libomp (#136)
Users wishing to use openmp through `-fopenmp` currently can not do so,
because libomp is available only as a shared library.

Shared libraries are not exposed by this toolchain because there is limited
utility in doing so (only the sandboxed compile step will succeed, and because
the lib is missing from runfiles, the executable won't run).

The proper way to depend on openmp will be the same as depending on other
shared libraries, i.e. either depend on the host/sysroot to provide it, or provide
it through `deps` using perhaps a `cc_import`.

While it was technically possible to create a `cc_import` target for libomp, it
was a little awkward because of the different extensions between darwin and
linux. So we now provide a convenience target `@llvm_toolchain//:omp` and
include a unit test as a demo.

Co-authored-by: Siddhartha Bagaria <sbagaria@grailbio.com>