Remove lpthread and ldl from link_libs
The original use case of #103 is better satisfied through
the users specifying `link_libs = {"": ["-lpthread", "-ldl"]}` in the
repository rule if they need it globally. Else they can specify it on a
per-target basis as `linkopt`.

We have started supporting the use of stdlib implementations that are
not libc++ and therefore do not require lpthread and ldl. I think the
historic availability of these libraries was by coincidence and we can
expect users to manually specify them if they need. This is also the
default with the LLVM toolchain outside of bazel.