Re-enable migration test
As of bazel 5.0, --all_incompatible_changes flag is a no-op. So we
should switch to using bazelisk for checking any migration issues.

https://github.com/bazelbuild/bazel/issues/13892

Because bazelisk does not allow specifying flag overrides on the command
line, we can not override the known failing incompatible flags, and so
our migration test will now fail. This is not ideal, but is the only
option going forward.