WIP: Add clang-cl windows cross compilation
Hi, this is going to be a little insane and a little strange but I would
like to get a read on the following as I refine it towards a potential
addition to these rules.

The following code allows for cross compiling windows (with MSVC ABI
rather than GNU, and I hope to adapt it to cope with clang on windows
too.

If this is something wanted I can polish these rules further rather than
fork and adapt.

This is very much an initial read, so there are some hacks, a hardcoded
path I could use help with and probably constructions that I am sure we
want to change.