Do not depend on gcc or libgcc install
We needed a gcc install in our setup because we depended on it for crt
sections not realizing that these are also shipped with more recent
versions of LLVM. See https://reviews.llvm.org/D28791 for more details.

There are situations when a user might prefer the libgcc implementation
but that should be controllable through options given to the clang
frontend, and we should make the LLVM crt object files available within
the sandbox anyway.

Container based tests have been changed to not install gcc anymore and
they continue to pass with this change. Inspecting the logs makes it
clear that the link step is picking clang_rt.crtbegin.o and
clang_rt.crtend.o from the sandbox.

Co-authored-by: Boleyn Su <boleyn.su@gmail.com>