Remove migration tests
These tests have not been useful for many past bazel versions, and
mostly not reliable. As bazel has become more stable, we expect
migration issues to be fewer. We do run tests with our minimum supported
bazel version and with the latest version of bazel. So a breakage with a
new release of bazel will be caught.