Update LLVM version in tests to 15
External tests still need to be on 14, because of crashes in some tests
with LLVM 15.

Suse tests work with LLVM 13 now, so they have been re-enabled
(fixes #143).

For macOS, Intel binaries are available only for 15.0.7, but for Linux,
binaries are available only for 15.0.6. So we now include a mechanism to
provide different version strings for different {os, arch} pair.