Re-enable omp_test on macOS
The issue was that depending on the libopenmp.dylib was adding the
toolchain library directory to the library search path, which is not
there by default. Having that directory on the search path was causing
libc++ and libc++abi to be linked dynamically from the toolchain but
those libraries are not present in the runfiles for the test. Linking
them statically was also not helpful because some system libraries load
them dynamically and could conflict with our static linked versions.
Ensuring that these are dynamic linked from the system sysroot is the
safest way.

Closes #179.