Add support for `layering_check` (#246)
* Add support for `layering_check`

The general support for this feature is inherited from the
Bazel-provided Unix C++ toolchain, except that a module map for the
toolchain and sysroot headers has to be supplied to the `cc_toolchain`.

Requires an update of abseil-cpp to make it compatible with
`layering_check`. Some other deps also needed to be updated.

Also fixes some nits related to Bazel 7 upgrade.