Allow debian to use LLVM 18 (#394)
No ubuntu 20.04 bundles are produced for LLVM 18, thus the check binding
to ubuntu 20.04 fails and complains there are no versions possible. Just
going by the ubuntu and debian release timelines debian >=9 should be
compatible with ubuntu 18 but >=8 might also work. For now not breaking
the existing >=9 check just using 18.04 when LLVM 18 is requested.