ci: update `libtinfo5` installation for Ubuntu 24.04
As per https://github.com/actions/runner-images/issues/10636
`ubuntu-latest` is now Ubuntu 24.04 instead of 22.04 which [does not]
have `libtinfo5` in the default PPAs

[does not]: https://packages.ubuntu.com/search?keywords=libtinfo

Unfortunately all of the LLVM binary releases still for Ubuntu still
target 18.04 and contain binaries linked against `libtinfo5` — bumping
the version of LLVM we use in our tests' toolchains will not solve this
problem.

Pinning CI to use Ubuntu 22.04 is an option; however this PR installs
`libtinfo5` on the Ubuntu 24.04 runners manually instead.