Fix shfmt config to disable Bash code simplification (#411)
I'm @ormandi's colleague and was following #409. It seems the formatter
wanting to remove quotes is due to Trunk by default using the `-s` flag
for `shfmt` to simplify code. Either there is a bug in `shfmt` that
makes it overly aggressive in removing quotes or what it's doing is
actually correct, but I'm not enough of a shell expert to tell.

This PR sidesteps the issue by customizes the Trunk config to avoid
passing `-s`.