Add dynamic-stdc++ (#400)
Depending on your project statically linking libstdc++ from your sysroot
might not be viable. For example if you rely on other shared libraries
that link the shared version, linking the static version to a binary
might result in ODR violations. This new `dynamic-stdc++` sets the
compile flags correctly but doesn't force the static version, so the
sysroot's shared version will be preferred.