Remove -fdebug-prefix-map for toolchain
The relative paths to toolchain components in binaries does not affect
hermeticity. It's possible these affect debugging instead, in which case
remapping them requires users to add reverse maps in their debugger
config. Instead leaving the paths as is makes these "just work" if users
have an external/ symlink in the root of their repo (which is common for
compile_commands.json as well).