support extra flags
If you just want to add a single flag (that is per platform) and leave the rest of the default
flags, you have to copy out all the existing flags and add it to what
you have. Instead support just adding an extra flag that is appended to
the default flags.