Make invocation of bash more Posix conformant.
Hard-coding the location of the shell to be in /bin/bash
is not always working as not all Posix-conformant operating
systems (such as NixOS) provide the path /bin/bash; so shell
scripts won't be able to execute.

Instead, use the `/usr/bin/env bash` idiom to reliably
find the shell to execute.

This fix is only applied to the critical scripts that might
be executed on the users' machine; the *_test*sh scripts
are left as-is.