Create sanitized response file instead of expanding arguments
The cc_wrapper.sh scripts inspects response files and sanitizes each
line within them.  How the logic is structured currently causes all of
these arguments to the be directly used later in the script.  This can
run into argument overflows.  It can also run into problems where
quoting in response files and from args is treated differently.

This addresses some of the comments in #430, namely cleaning up files
and not overwriting the input files.  I have not tried to apply a
threshold here as that is best left up to the tooling that previously
decided to use response files, which can cause difficult to debug
changes in behavior.

This conflicts with #479, but is trying to address the same underlying
problem of quoted arguments in response files as the result of golang's
link actions.

Fixes #421