Verify the current resolution (#490)
* Split `llvm_release_name` into two:
* `llvm_release_name_host_info`: Used to verify the resolution. It does
not take a rctx.
* `llvm_release_name_context`: Will fail if resolution from rctx fails.
* Change `dist_version_arch` to `host_info` which returns a struct, so
the result can easily be extended and only one call is needed.

As can be seen in `toolchain/internal/llvm_distributions.golden.sel.txt`
there are a number of decisions taken in mode `auto` that lead to
failures in predicting a correct LLVM distribution. Those we should fix.