Support extra toolchain flags (#554)
If you just want to add a single flag (that is per platform) and leave
the rest of the default
flags, you have to copy out all the existing flags and add it to what
you have. Instead support just adding an extra flag that is appended to
the default flags.

This is a resubmit of #452 by @rockwotj with minor fixes in
documentation.

---------

Co-authored-by: Tyler Rockwood <rockwood@redpanda.com>