Update rules_cc to 0.2.14. (#605)
With newer versions of `rules_cc` for WORKSPACE setups we need the
following new flag that has to be put into the relevant .bazelrc files:
`common --repositories_without_autoloads=cc_compatibility_proxy`.

We also need a main `WORKSPACE.bzlmod` in order for `--enable_bzlmod
--enable_workspace` to work (false/true and true/false) already worked
but (true/true) failed in Bazel 8 before this.